Rename directory to package name
