+++ morpheus/framework/resources/ostype/example.tf.tmpl
