+++ morpheus/framework/resources/ostype/os_type_example.go
