+++ morpheus/framework/resources/ostype/resource_delete.go
