+++ morpheus/framework/resources/ostype/resource_read.go
